requirements.txt
numpy
pandas
fastapi
uvicorn
sqlalchemy
scikit-learn
python-dotenv
psycopg2-binary